+++ driver/aws/dynamojournal/store.go
@@ -4,13 +4,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"sync"
-	"sync/atomic"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/dogmatiq/persistencekit/driver/aws/internal/awsx"
+	"github.com/dogmatiq/persistencekit/internal/syncx"
 	"github.com/dogmatiq/persistencekit/journal"
 )
 
@@ -34,8 +33,7 @@ type BinaryStore struct {
 	// options before the request is sent.
 	OnRequest func(any) []func(*dynamodb.Options)
 
-	created  atomic.Bool
-	createdM sync.Mutex
+	create syncx.SucceedOnce
 }
 
 const (
@@ -46,6 +44,10 @@ const (
 
 // Open returns the journal with the given name.
 func (s *BinaryStore) Open(ctx context.Context, name string) (journal.BinaryJournal, error) {
+	if s.Table == "" {
+		panic("table name must not be empty")
+	}
+
 	if err := s.createTable(ctx); err != nil {
 		return nil, err
 	}
@@ -126,50 +128,40 @@ func (s *BinaryStore) Open(ctx context.Context, name string) (journal.BinaryJour
 }
 
 func (s *BinaryStore) createTable(ctx context.Context) error {
-	if s.created.Load() {
-		return nil
-	}
-
-	s.createdM.Lock()
-	defer s.createdM.Unlock()
-
-	if s.created.Load() {
-		return nil
-	}
-
-	if _, err := awsx.Do(
-		ctx,
-		s.Client.CreateTable,
-		s.OnRequest,
-		&dynamodb.CreateTableInput{
-			TableName: aws.String(s.Table),
-			AttributeDefinitions: []types.AttributeDefinition{
-				{
-					AttributeName: aws.String(nameAttr),
-					AttributeType: types.ScalarAttributeTypeS,
+	return s.create.Do(
+		func() error {
+			if _, err := awsx.Do(
+				ctx,
+				s.Client.CreateTable,
+				s.OnRequest,
+				&dynamodb.CreateTableInput{
+					TableName: aws.String(s.Table),
+					AttributeDefinitions: []types.AttributeDefinition{
+						{
+							AttributeName: aws.String(nameAttr),
+							AttributeType: types.ScalarAttributeTypeS,
+						},
+						{
+							AttributeName: aws.String(positionAttr),
+							AttributeType: types.ScalarAttributeTypeN,
+						},
+					},
+					KeySchema: []types.KeySchemaElement{
+						{
+							AttributeName: aws.String(nameAttr),
+							KeyType:       types.KeyTypeHash,
+						},
+						{
+							AttributeName: aws.String(positionAttr),
+							KeyType:       types.KeyTypeRange,
+						},
+					},
+					BillingMode: types.BillingModePayPerRequest,
 				},
-				{
-					AttributeName: aws.String(positionAttr),
-					AttributeType: types.ScalarAttributeTypeN,
-				},
-			},
-			KeySchema: []types.KeySchemaElement{
-				{
-					AttributeName: aws.String(nameAttr),
-					KeyType:       types.KeyTypeHash,
-				},
-				{
-					AttributeName: aws.String(positionAttr),
-					KeyType:       types.KeyTypeRange,
-				},
-			},
-			BillingMode: types.BillingModePayPerRequest,
+			); err != nil && !errors.As(err, new(*types.ResourceInUseException)) {
+				return fmt.Errorf("unable to create DynamoDB table: %w", err)
+			}
+			return nil
 		},
-	); err != nil && !errors.As(err, new(*types.ResourceInUseException)) {
-		return fmt.Errorf("unable to create DynamoDB table: %w", err)
-	}
-
-	s.created.Store(true)
-
-	return nil
+	)
 }