+++ driver/aws/dynamokv/store_test.go
@@ -41,13 +41,6 @@ func setup(t testing.TB) (*dynamodb.Client, string) {
 	client := dynamox.NewTestClient(t)
 	table := "kvstore"
 
-	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-	defer cancel()
-
-	if err := CreateTable(ctx, client, table); err != nil {
-		t.Fatal(err)
-	}
-
 	t.Cleanup(func() {
 		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 		defer cancel()